(BSR)[PRO] refactor: remove music types from offer individual context
