+++ pro/src/utils/individualApiFactories.ts
@@ -181,23 +181,6 @@ export const individualOfferContextValuesFactory = (
     offer,
     categories: [],
     subCategories: [],
-    musicTypes: [
-      {
-        gtl_id: '01000000',
-        label: 'Musique Classique',
-        canBeEvent: true,
-      },
-      {
-        gtl_id: '07000000',
-        label: 'Metal',
-        canBeEvent: true,
-      },
-      {
-        gtl_id: '15000000',
-        label: 'Vidéos musicales',
-        canBeEvent: false,
-      },
-    ],
     ...customIndividualOfferContextValues,
   }
 }
